requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
aiohttp==3.9.1
pydantic==2.5.2
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.13.1
redis==5.0.1
celery==5.3.4

# AI/ML dependencies
openai==1.6.1
openai-whisper==20231117
torch==2.1.2
torchaudio==2.1.2
transformers==4.36.2
sentence-transformers==2.2.2

# Audio/Video processing
librosa==0.10.1
moviepy==1.0.3
opencv-python==4.8.1.78
ffmpeg-python==0.2.0
Pillow==10.1.0

# Data processing
numpy==1.24.3
pandas==2.0.3
scipy==1.11.4
textstat==0.7.3

# Utilities
requests==2.31.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-slugify==8.0.1
click==8.1.7

# Monitoring & logging
structlog==23.2.0
sentry-sdk==1.38.0
prometheus-client==0.19.0
